Add special handling for nil columns
Previously when the column value was `nil` then it generated a column
with empty string in that column so the column was not actually
blank/empty.

Now when the column is `nil` then it will add this column to generated
row XML without any content or content type information.

As a result if Google Sheets are used then the formula `=ISBLANK(A1)`
will return `TRUE`, previously it returned `FALSE`.

JRVS-306